refactor: add new core module structure
