+++ lib/src/core/core.dart
@@ -5,6 +5,6 @@ export 'errors/auto_cache_error.dart';
 
 export 'services/directory_service/directory_provider_service.dart';
 
-export 'utils/either.dart';
+export 'functional/either.dart';
 
 export 'extensions/type_extensions.dart';
