+++ lib/src/core/functional/either.dart
