+++ lib/src/core/infrastructure/protocols/cache_response.dart
