+++ lib/src/core/infrastructure/protocols/enums/cache_response_status.dart
